refactored the js code
