+++ src/index.html
@@ -28,7 +28,7 @@
         suscipit enim
         in ut!
     </p>
-    <script src="index.js"></script>
+    <script src="index.js" type="module"></script>
 
 </body>
