+++ src/index.js
@@ -1,10 +1,13 @@
+import {addFootnote, removeAllFootnotes, toggleFootnote, removeFootnote} from './actions.js'
+import {distance} from './utils.js'
+
 let content = {}
 let ftns = document.querySelectorAll("t-ftn") //selects all footnotes
 
 //removes all footnotes except the one that is opened
 window.addEventListener("click", (e) => removeAllFootnotes(e.target));
 
-for( i of ftns){
+for(let i of ftns){
     content[i] = i.innerHTML
 
     // option grabbing
@@ -21,17 +24,6 @@ for( i of ftns){
     
 }
 
-function distance(tdiv, bdiv) {
-    const trect = tdiv.getBoundingClientRect();
-    const brect = bdiv.getBoundingClientRect();
-    const bleftx = brect.left;
-    const blefty = brect.top;
-    const tcenterx = trect.left + trect.width / 2;
-    const tbottomy = trect.bottom;
-    const distance = Math.hypot(bleftx - tcenterx, tbottomy - blefty);
-
-    return distance;
-}
 
 // Create a class for the element
 class Footnote extends HTMLElement {
@@ -88,36 +80,3 @@ class Footnote extends HTMLElement {
   }
   
 customElements.define("t-ftn", Footnote);
-  
-
-// toggles visibility when the note is clicked
-function toggleFootnote(){
-    removeAllFootnotes(this);
-    var vis = this.shadowRoot.children[1].style.visibility;
-    if(vis == 'hidden'){
-        this.shadowRoot.children[1].style.visibility = 'visible';
-    }else{
-        this.shadowRoot.children[1].style.visibility = 'hidden';
-    }
-}
-
-// adds visibility to clicked footnote
-function addFootnote(){
-    removeAllFootnotes();
-    this.shadowRoot.children[1].style.visibility = 'visible';
-}
-
-// removes visibility from clicked footnote
-function removeFootnote(){
-    this.shadowRoot.children[1].style.visibility = 'hidden';
-}
-
-// removes visibility from all footnotes except for current note (if there is one)
-// curr = current note
-function removeAllFootnotes(curr){
-    document.querySelectorAll('t-ftn').forEach(ftn => {
-        if(ftn !== curr){
-            ftn.shadowRoot.children[1].style.visibility = 'hidden';
-        }
-    })
-}
